Fix logger bugs
